analyze done.........................

+++ dma-back/src/main/kotlin/linKav20/github/com/entities/results/analyze/models/AnalyzeModel.kt
@@ -3,4 +3,4 @@ package linKav20.github.com.entities.results.analyze.models
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class AnalyzeModel(var parent: String, var category: String, var stats: Long);
+data class AnalyzeModel(var parent: String, var category: String, var stats: Int);
